+++ @ecomplus/storefront-app/src/views/Account.vue
@@ -11,7 +11,7 @@
         <ec-account-form :customer.sync="customer"/>
       </template>
       <template #account-addresses>
-        <ec-addresses :customer.sync="customer"/>
+        <account-addresses :customer.sync="customer"/>
       </template>
       <template #orders-list>
         <ec-orders-list