+++ @ecomplus/storefront-components/src/TheCart.vue
@@ -1,0 +1,3 @@
+<script src="./js/TheCart.js"></script>
+<template lang="html" src="./html/TheCart.html"></template>
+<style lang="scss" src="./scss/TheCart.scss"></style>