+++ @ecomplus/storefront-components/src/js/TheCart.js
@@ -1,26 +1,24 @@
 import {
-  $ecomConfig,
-  i18n,
-  formatMoney
-} from '@ecomplus/utils'
-
-import ecomCart from '@ecomplus/shopping-cart'
-import baseModulesRequestData from '../../lib/base-modules-request-data'
-import APrices from '#components/APrices.vue'
-import CartItem from '#components/CartItem.vue'
-import DiscountApplier from '#components/DiscountApplier.vue'
-import ShippingCalculator from '#components/ShippingCalculator.vue'
-import RecommendedItems from '#components/RecommendedItems.vue'
-
-import {
   i19checkout,
   i19continueShopping,
   i19discount,
   i19emptyCart
 } from '@ecomplus/i18n'
 
+import {
+  i18n,
+  formatMoney
+} from '@ecomplus/utils'
+
+import ecomCart from '@ecomplus/shopping-cart'
+import APrices from './../APrices.vue'
+import CartItem from './../CartItem.vue'
+import DiscountApplier from './../DiscountApplier.vue'
+import ShippingCalculator from './../ShippingCalculator.vue'
+import RecommendedItems from './../RecommendedItems.vue'
+
 export default {
-  name: 'EcCart',
+  name: 'TheCart',
 
   components: {
     APrices,
@@ -31,30 +29,32 @@ export default {
   },
 
   props: {
-    ecomCart: {
+    amount: {
       type: Object,
-      default: () => ecomCart
-    },
-    lang: {
-      type: String,
-      default: $ecomConfig.get('lang')
+      default () {
+        return {}
+      }
     },
     checkoutUrl: {
       type: String,
       default: '/app/#/checkout'
     },
-    amount: {
-      type: Object,
-      default: () => {}
-    },
     zipCode: String,
-    discountCoupon: String
+    discountCoupon: String,
+    modulesPayload: Object,
+    ecomCart: {
+      type: Object,
+      default () {
+        return ecomCart
+      }
+    }
   },
 
   data () {
     return {
       localZipCode: this.zipCode,
-      canApplyDiscount: false
+      canApplyDiscount: false,
+      isCouponApplied: false
     }
   },
 
@@ -63,7 +63,6 @@ export default {
     i19continueShopping: () => i18n(i19continueShopping),
     i19discount: () => i18n(i19discount),
     i19emptyCart: () => i18n(i19emptyCart),
-    modulesPayload: () => baseModulesRequestData,
 
     cart () {
       return this.ecomCart.data
@@ -74,7 +73,7 @@ export default {
         return this.discountCoupon
       },
       set (couponCode) {
-        this.$emit('update:discountCoupon', couponCode)
+        this.$emit('update:discount-coupon', couponCode)
       }
     }
   },
@@ -83,16 +82,29 @@ export default {
     formatMoney,
 
     selectShippingService (service) {
-      this.$emit('shippingService', service)
+      this.$emit('select-shipping', service)
       this.$nextTick(() => {
         this.canApplyDiscount = true
+      })
+    },
+
+    setDiscountRule (discountRule) {
+      this.$emit('set-discount-rule', discountRule)
+      this.$nextTick(() => {
+        this.isCouponApplied = Boolean(this.discountCoupon && this.amount.discount)
       })
     }
   },
 
   watch: {
     localZipCode (zipCode) {
-      this.$emit('update:zipCode', zipCode)
+      this.$emit('update:zip-code', zipCode)
+    },
+
+    canApplyDiscount (canApplyDiscount) {
+      if (!canApplyDiscount) {
+        this.isCouponApplied = false
+      }
     }
   },