+++ @ecomplus/storefront-template/CHANGELOG.md
@@ -3,6 +3,29 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.0.0-beta.43](https://github.com/ecomplus/storefront/compare/@ecomplus/storefront-template@2.0.0-beta.42...@ecomplus/storefront-template@2.0.0-beta.43) (2020-06-30)
+
+
+### Bug Fixes
+
+* **deps:** update @ecomplus/shopping-cart to v2.1.2 ([d4fb6b0](https://github.com/ecomplus/storefront/commit/d4fb6b032b82519f7e3b433fd6faad3715746fbf))
+* **deps:** update @ecomplus/shopping-cart to v2.1.3 ([0a71d3b](https://github.com/ecomplus/storefront/commit/0a71d3b161b8f2b4258ec72df0772a5ac9e8b3c0))
+
+
+
+
+
+# [2.0.0-beta.42](https://github.com/ecomplus/storefront/compare/@ecomplus/storefront-template@2.0.0-beta.41...@ecomplus/storefront-template@2.0.0-beta.42) (2020-06-25)
+
+
+### Features
+
+* **theme:** add new 'niche baby' theme ([#245](https://github.com/ecomplus/storefront/issues/245)) ([8d034b4](https://github.com/ecomplus/storefront/commit/8d034b461b21f7135a2b0cb67b2a512713a93392))
+
+
+
+
+
 # [2.0.0-beta.41](https://github.com/ecomplus/storefront/compare/@ecomplus/storefront-template@2.0.0-beta.40...@ecomplus/storefront-template@2.0.0-beta.41) (2020-06-21)
 
 **Note:** Version bump only for package @ecomplus/storefront-template